q2_multimodal_qa_agent/requirements.txt
streamlit
requests
python-dotenv
Pillow
matplotlib
google-generativeai>=0.5.0 